# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: dnsaur API
  version: "1.0"
  description: >
    REST API for the dnsaur DNS/DHCP server. All paths are prefixed
    /api/v1. Authenticate with either the dnsaur_session cookie (set by
    POST /auth/login) or an Authorization: Bearer <token> API token
    (see /tokens). API tokens may have scope "read" (GET/HEAD only) or
    "write" (all methods); the session cookie always has full access.
    A read-scoped token used for a POST/PUT/PATCH/DELETE request is
    rejected with 403 {"error": "read-only token"} — see the 403
    response documented on each mutating operation below. Scope is
    enforced by method, with one exception on the read side: a
    read-scoped token gets TSIG secrets redacted (see the TSIGKey
    schema's secret_redacted field), because a GET that returns a
    signing credential in the clear is not something "read-only"
    should cover. The Authorization scheme name is matched
    case-insensitively per RFC 9110 §11.1.
    There is no separate "upstreams" resource: upstream DNS servers are
    configured via the "upstreams" key in /settings (a comma-separated
    list), alongside "upstream.strategy". Each entry is plain by default
    ("1.1.1.1:53") or, with a scheme, DNS-over-TLS
    ("tls://1.1.1.1:853#cloudflare-dns.com") or DNS-over-HTTPS
    ("https://1.1.1.1/dns-query#cloudflare-dns.com") — see the "upstreams"
    key under PUT /settings for the full grammar.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dnsaur_session
      description: Session cookie set by POST /auth/login. HttpOnly, SameSite=Strict.
    bearerAuth:
      type: http
      scheme: bearer
      description: API token minted via POST /tokens, sent as "Authorization Bearer <token>".
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    TSIGKey:
      type: object
      description: >
        A TSIG key (RFC 8945) authenticates a zone transfer between this
        server and a peer -- the credential a secondary presents to pull a
        zone, or this server presents when it is the one transferring.
      properties:
        id: { type: integer, format: int64 }
        name:
          type: string
          description: >
            The TSIG RR's owner name, canonical form (lowercase, trailing
            dot) -- e.g. "xfer.e412.in.". Normalised with the same rule on
            every write, so a signed message's owner name can never miss the
            key it names on case or a missing dot. Must be unique: it is the
            only thing a signed message carries to say which key signed it.
        algorithm:
          type: string
          enum: [hmac-sha1., hmac-sha224., hmac-sha256., hmac-sha384., hmac-sha512.]
          description: >
            miekg/dns's own algorithm names, trailing dot included.
            hmac-md5.sig-alg.reg.int. is a real RFC 8945 registry name but is
            rejected here: the library removed MD5 support from its signer,
            so a key created with it would parse and store fine and then
            fail to sign the first transfer it was needed for.
        secret:
          type: string
          description: >
            Base64, and returned in the clear on every read to a session or
            a write-scoped token -- not hidden the way a freshly minted API
            token is. An API token only ever needs to be shown once, because
            only the holder who requested it uses it again. A TSIG secret is
            different: the same value has to be configured identically on
            the peer (BIND's key{} clause, Technitium's transfer settings),
            so a caller must be able to come back and read it whenever that
            peer needs re-pairing. Hiding it after creation would make the
            feature unusable without direct database access.
            Empty for a read-scoped token -- see secret_redacted.
        secret_redacted:
          type: boolean
          description: >
            True when secret was withheld because the caller presented a
            read-scoped API token. Scope is enforced by method (a read token
            may call any GET), and this is the only credential the API
            returns in the clear on a GET, so a token minted to let a
            dashboard look around would otherwise hand over every signing
            key the server holds. Always present, so an empty secret you may
            not see is distinguishable from one that was never set.
        created_at: { type: integer, format: int64, description: unix ms }
...
  responses:
    ReadOnlyToken:
      description: >
        read-only token used for a write operation, or a cookie-authenticated
        write whose Sec-Fetch-Site is cross-site ({"error": "cross-site
        request"}) -- CSRF defence in depth behind SameSite=Strict. Bearer
        requests are never checked for the latter.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /tsig-keys:
    get:
      tags: [TSIG Keys]
      summary: List TSIG keys
      description: >
        Includes each key's secret -- see the TSIGKey schema's secret field
        for why.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TSIGKey" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      tags: [TSIG Keys]
      summary: Create a TSIG key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, algorithm, secret]
              properties:
                name: { type: string }
                algorithm:
                  type: string
                  enum: [hmac-sha1., hmac-sha224., hmac-sha256., hmac-sha384., hmac-sha512.]
                secret: { type: string, description: base64 }
      responses:
        "201":
          description: Created
          headers:
            Location:
              description: URL of the key this request created
              schema: { type: string, example: /api/v1/tsig-keys/2 }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TSIGKey"
        "400":
          description: >
            name is not a valid domain name, algorithm is not one of the
            five miekg/dns can sign with, or secret is not valid base64
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
        "409":
          description: a TSIG key with that name already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /tsig-keys/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [TSIG Keys]
      summary: Get a TSIG key
      description: >
        Includes the secret -- see the TSIGKey schema's secret field for why.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TSIGKey" }
        "400":
          description: bad id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    put:
      tags: [TSIG Keys]
      summary: Replace a TSIG key
      description: >
        Full replacement, not a partial patch -- name, algorithm and secret
        are all required, the same as create.


        A *rename* is refused with 409 "resource in use" while any zone's
        allow_transfer or notify_to names this key by name: renaming it
        would leave that zone naming a key that no longer exists, and every
        signed transfer or NOTIFY under it refused with nothing to say why
        -- the same reason DELETE is refused. Everything else stays
        editable on a referenced key, rotating the secret included, and a
        key referenced only by a zone's tsig_key_id can be renamed freely,
        since that reference is by id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, algorithm, secret]
              properties:
                name: { type: string }
                algorithm:
                  type: string
                  enum: [hmac-sha1., hmac-sha224., hmac-sha256., hmac-sha384., hmac-sha512.]
                secret: { type: string, description: base64 }
      responses:
        "204":
          description: No Content
        "400":
          description: >
            bad id, invalid json, name is not a valid domain name,
            algorithm is not one of the five miekg/dns can sign with, or
            secret is not valid base64
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: >
            "a TSIG key with that name already exists", or "resource in
            use" when the write renames a key a zone names in
            allow_transfer or notify_to
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [TSIG Keys]
      summary: Delete a TSIG key
      description: >
        Refused with 409 while any zone names this key — by tsig_key_id, or
        by `key:` in allow_transfer or notify_to. Deleting it would leave
        that zone unable to authenticate its transfers, with nothing on the
        zone to say why — so the key has to be released from every zone
        that uses it first. PUT applies the name half of the same guard.
      responses:
        "204":
          description: No Content
        "400":
          description: bad id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: a zone still references this key ("resource in use")
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...